Multimodulize project
part of vaadin/flow-component-base#117
